Possible fix for control selection bug
Trying to fix a reported bug that I've been unable to reproduce where multiple handles are being selected when just one is clicked. I assume this is because they're somehow being assigned the same id, but not sure how this is possible. Hopefully this hacky fix will work until I can figure out what's going on.