Optimized VertexPath.CalculateClosestPointOnPathData method.
On my PC, with the provided test scene "GetClosestPointOnPath Test" the profiling results are:
- Old implementation: 5.58ms - 14.62ms
- New implementation: 0.25ms - 0.40ms

The idea for the optimization is to do the math in "VertexPath local-space" rather than world-space.
This optimization thus avoids to transform hundrets or thousands of path positions to world-space.
It only applies two transformations instead: One to transform the incoming worldPoint into local-space and then a transform of the computed result from local- to world-space.